# Conflicts:
#	indexer/modules/custom/eigen_layer/models/af_eigen_layer_address_current.py
#	indexer/modules/custom/eigen_layer/models/af_eigen_layer_records.py

+++ indexer/modules/custom/stats/models/daily_bridge_transactions_stats.py
@@ -1,4 +1,4 @@
-from sqlalchemy import NUMERIC, BigInteger, Column
+from sqlalchemy import Column
 from sqlalchemy.dialects.postgresql import BIGINT, DATE
 
 from common.models import HemeraModel
@@ -9,5 +9,5 @@ class DailyBridgeTransactionsAggregates(HemeraModel):
     __tablename__ = "af_stats_na_daily_bridge_transactions"
 
     block_date = Column(DATE, primary_key=True)
-    deposit_cnt = Column(BigInteger)
-    withdraw_cnt = Column(BigInteger)
+    deposit_cnt = Column(BIGINT)
+    withdraw_cnt = Column(BIGINT)
